graph_sentiment/flask_requirements.txt
Flask 
numpy 
pandas 
gunicorn 
psycopg2-binary 
plotly
flasgger
flask-cors
